1.加入了libevent-2.1.8-stable 2.加入了libhttp用以假设web服务器 3.为libCScriptEng增加了更多的库 4.编写了hmhttptest模块供libhttp使用，使得可以通过浏览器访问文件系统 5.在hmhttptest模块中允许以一种类似于php的方式加载mzchtml文件来提供动态服务器页面生成功能

+++ core/CScriptEng/functionStatement.cpp
@@ -174,7 +174,7 @@ namespace compiler {
 			break;
 
 			// 对于压入的参数，无需理会
-		case -1:
+		case (uint32_t)-1:
 			break;
 
 		default:
